// libs/orders/src/lib/models/order.ts
// import { Product }  from '@cakefairy/products';

export class Order {
  id?: string;
  product?: any;
  quantity?: number;  
  apartment?: string;  
  city?: string;
  street?: string;
  zip?: string;
  district?: string;  
  phone?: string;
  status?: number;
  totalPrice?: number;
  user?: any;
  dateOrdered?: string;  
}